Module uic_attribute_store_rs::attribute_store

Enums

AttributeStoreTimeOutError

Functions

add_node

Add node to the attribute store, returns Result with node_id or Err if node_id is ATTRIBUTE_STORE_INVALID_NODE

attribute_store_get_node_attribute_value

attribute_store_set_node_attribute_value

get_or_add_node

Get or add node (and set value) of an attribute store node If value is Some it will do an attribute_store_get_node_child_by_value(), if not it will do attribute_store_get_node_child_by_type If set_value is Some it will set the value in the attribute that is get or created. Returns the node_id of the created (og get) node.

init

log

read_attribute

Read attribute from the attribute store in a type-safe manner

set_callback_on_timeout

teardown

write_attribute

Write attribute value to the attribute store. param value requires serde::Serialize implementation!