Function uic_attribute_store_rs::attribute_store_set_reported_string

pub unsafe extern "C" fn attribute_store_set_reported_string(
    node: attribute_store_node_t, 
    string: *const c_char
) -> sl_status_t

@brief Safely copies a String in the Attribute Store

@param node The Attribute store node under which the value must be written @param string C char array containing a String

@returns sl_status_t returned by @ref attribute_store_set_node_attribute_value